Split file-node test wrappers
Move SQLite and Postgres file-node test entry points into sibling modules
so `src/db/tests/mod.rs` stays below the repository file-size limit.

Keep the shared fixture and non-file-node database smoke tests in the
parent module while preserving the existing scenario bodies and harness
calls.
